Share listener lifecycle across Falcon services
