+++ lib/falcon/service/server.rb
@@ -7,6 +7,7 @@
 require "async/service/managed/service"
 require "async/http/endpoint"
 
+require_relative "../listener"
 require_relative "../server"
 
 module Falcon
@@ -18,7 +19,21 @@ module Falcon
 			def initialize(...)
 				super
 				
-				@bound_endpoint = nil
+				@listener = nil
+			end
+			
+			# Build a listener from a configured and bound endpoint.
+			# @parameter evaluator [Environment::Evaluator] The environment evaluator.
+			# @parameter endpoint [Async::HTTP::Endpoint] The configured endpoint.
+			# @parameter bound_endpoint [IO::Endpoint::BoundEndpoint] The bound endpoint.
+			# @returns [Falcon::Listener] The bound listener.
+			def make_listener(evaluator, endpoint, bound_endpoint)
+				Listener.new(
+					name: evaluator.name,
+					scheme: endpoint.scheme,
+					protocols: endpoint.protocol.names,
+					endpoint: bound_endpoint,
+				)
 			end
 			
 			# Bind the endpoint used by each server worker.
@@ -26,7 +41,8 @@ module Falcon
 				@endpoint = @evaluator.endpoint
 				
 				Sync do
-					@bound_endpoint = @endpoint.bound
+					bound_endpoint = @endpoint.bound
+					@listener = make_listener(@evaluator, @endpoint, bound_endpoint)
 				end
 				
 				Console.info(self){"Starting #{self.name} on #{@endpoint}"}
@@ -39,20 +55,63 @@ module Falcon
 				super
 			end
 			
+			# Yield the listener used by a server worker.
+			# @parameter evaluator [Environment::Evaluator] The environment evaluator.
+			# @yields {|listener| ...} The listener used by the worker.
+			# 	@parameter listener [Falcon::Listener] The bound listener.
+			def with_listener(evaluator)
+				yield @listener
+			end
+			
+			# Setup the service into the specified container.
+			# @parameter container [Async::Container] The container to configure.
+			def setup(container)
+				container_options = @evaluator.container_options
+				health_check_timeout = container_options[:health_check_timeout]
+				
+				container.run(**container_options) do |instance|
+					clock = Async::Clock.start
+					evaluator = self.environment.evaluator
+					
+					with_listener(evaluator) do |listener|
+						Async do
+							server = nil
+							
+							health_checker(instance, health_check_timeout) do
+								if server
+									instance.name = format_title(evaluator, server)
+								end
+							end
+							
+							instance.status!("Preparing...")
+							evaluator.prepare_worker!(instance, listener: listener)
+							emit_prepared(instance, clock)
+							
+							instance.status!("Running...")
+							server = run(instance, evaluator, listener)
+							instance.name = format_title(evaluator, server)
+							emit_running(instance, clock)
+							
+							instance.ready!
+						end
+					end
+				end
+			end
+			
 			# Run the service logic.
 			#
 			# @parameter instance [Object] The container instance.
 			# @parameter evaluator [Environment::Evaluator] The environment evaluator.
-			# @parameter bound_endpoint [IO::Endpoint] The endpoint bound by this worker.
+			# @parameter listener [Falcon::Listener] The listener used by this worker.
 			# @returns [Falcon::Server] The server instance.
-			def run(instance, evaluator, bound_endpoint = @bound_endpoint)
+			def run(instance, evaluator, listener = @listener)
 				if evaluator.respond_to?(:make_supervised_worker)
 					Console.warn(self, "Async::Container::Supervisor is replaced by Async::Service::Supervisor, please update your service definition.")
 					
 					evaluator.make_supervised_worker(instance).run
 				end
 				
-				server = evaluator.make_server(bound_endpoint)
+				server = evaluator.make_server(listener.endpoint)
 				
 				Async do |task|
 					server.run
@@ -75,9 +134,9 @@ module Falcon
 			
 			# Close the bound endpoint.
 			def stop(...)
-				if @bound_endpoint
-					@bound_endpoint.close
-					@bound_endpoint = nil
+				if @listener
+					@listener.endpoint.close
+					@listener = nil
 				end
 				
 				@endpoint = nil
